// src/app/enums/action.enum.ts
export enum ActionEnum {
    filter = "filter",
    view = "view",
    delete = "delete",
    update = "update",
    actions = "actions",
    changeMeal = "changeMeal",
    changeDeliveryDetails = "changeDeliveryDetails",
    changeDeliveryNotes = "changeDeliveryNotes",
    createSubscription = "createSubscription",
    changeStatus = "changeStatus",
}